// src/types/reports/reports.ts
export interface IReportsStore {
  label: string;
  value: string;
}
// ...
export interface IChartData {
  total: number;
  data: Array<{
    name: string;
    value: number;
  }>;
}